# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Mavi API
  version: 0.1.0
  description: |
    JSON CRUD for candidates, their profiles, roles and matches. Every request
    outside `/health` names a persona with the `X-Role` header (`talent`,
    `employer` or `ops`) and may identify itself with `X-Actor` (the candidate
    id for talent, an email for ops). A missing or unknown role is a 401; a
    known role calling an operation it may not use is a 403.

    `PUT` on candidates and roles replaces the fields you send and keeps the
    rest; an explicit `null` clears an optional field and is a 422 on a
    required one. `PUT …/profile` replaces the whole profile. Taxonomy fields
    (certifications, software) accept free text and are stored as canonical
    ids from infra/taxonomy.json; an unknown value is a 422.
servers:
  - url: http://localhost:8080
security:
  - role: []
...
paths:
...
  /candidates:
    post:
      tags: [candidates]
      operationId: createCandidate
      summary: Create a candidate
      description: "Talent-created candidates are marked `source: self` unless a source is given."
      x-roles: [talent, ops]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Candidate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    get:
      tags: [candidates]
      operationId: listCandidates
      summary: List candidates
      x-roles: [ops]
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/CandidateStatus" }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Candidate" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /candidates/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [candidates]
      operationId: getCandidate
      summary: Fetch a candidate
      description: Talent may only read their own record (`X-Actor` = id); another id is a 404.
      x-roles: [talent, ops]
      responses:
        "200":
          description: The candidate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Candidate" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [candidates]
      operationId: updateCandidate
      summary: Update a candidate
      description: Replaces the fields sent and keeps the rest. Talent may only update their own record.
      x-roles: [talent, ops]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Candidate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [candidates]
      operationId: deleteCandidate
      summary: Delete a candidate
      x-roles: [ops]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /candidates/{id}/profile:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [profiles]
      operationId: getProfile
      summary: Fetch a candidate's structured profile
      x-roles: [talent, ops]
      responses:
        "200":
          description: The profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [profiles]
      operationId: putProfile
      summary: Create or replace a candidate's profile
      description: |
        Replaces the whole profile. Omitted lists become empty and an omitted
        availability becomes `unknown`. Re-saving identical content keeps the
        stored embedding; a change to the text it was derived from clears it.
      x-roles: [talent, ops]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileInput" }
      responses:
        "200":
          description: Replaced an existing profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "201":
          description: Created the candidate's first profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [profiles]
      operationId: deleteProfile
      summary: Delete a candidate's profile
      x-roles: [ops]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /candidates/{id}/resume:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [profiles]
      operationId: uploadResume
      summary: Upload a resume and queue its parsing
      description: |
        Talent intake. The request body is the file itself (not multipart): a
        PDF, or a DOCX. What the file is is read from its content, not from
        `Content-Type`. Its text is extracted in the request and stored as the
        candidate's `resume_text`, and a `parse_resume` job is queued; the
        response is that job, whose `id` is what to poll. The worker sends the
        text to the AI service's `/parse-resume`, writes the result as the
        candidate's profile (replacing any profile already there) and embeds
        it. Poll `GET /candidates/{id}/resume/job` (or, as ops,
        `GET /jobs/{id}`) until `status` is `succeeded`, then read
        `GET /candidates/{id}/profile`.

        Uploading again while the job is still queued replaces the text and
        returns the same job. Talent may only upload to their own record.
      x-roles: [talent, ops]
      requestBody:
        required: true
        description: The file itself, at most 5 MB and 10 pages.
        content:
          application/pdf:
            schema: { type: string, format: binary }
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        "202":
          description: Text extracted and stored; the parse is queued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413":
          description: The file is larger than 5 MB.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "415":
          description: The file is neither a PDF nor a DOCX.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: The file is empty, damaged, password-protected, has no text, or has too many pages or too much text; `error` says which.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: The AI service, which reads the file, is unavailable. Nothing was stored; try again.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /candidates/{id}/resume/job:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [profiles]
      operationId: getResumeJob
      summary: Status of the candidate's most recent resume parse
      description: |
        The newest `parse_resume` job for the candidate: the same object
        `GET /jobs/{id}` returns, readable by the talent who uploaded. The
        profile is ready when `status` is `succeeded`; on `failed`,
        `last_error` says why. A 404 when no resume was ever uploaded.
      x-roles: [talent, ops]
      responses:
        "200":
          description: The job.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /candidates/{id}/availability:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [availability]
      operationId: getAvailability
      summary: Fetch what the candidate said about when and where they can work
      description: A 404 until the candidate has supplied it; such a candidate is excluded from matching.
      x-roles: [talent, ops]
      responses:
        "200":
          description: The candidate's availability.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkAvailability" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [availability]
      operationId: putAvailability
      summary: Set or edit the candidate's time zone, working hours, hours per week and start date
      description: |
        Availability and time-zone overlap are hard filters that a resume
        cannot answer, so the candidate supplies them, here. All four are
        required and the body replaces whatever was stored. Parsing a resume
        (again) never changes them. Talent may only set their own.
      x-roles: [talent, ops]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkAvailabilityInput" }
      responses:
        "200":
          description: Replaced what was stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkAvailability" }
        "201":
          description: Stored for the first time.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkAvailability" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  # ---------------------------------------------------------------------------
  # Roles
  # ---------------------------------------------------------------------------
  /roles:
    post:
      tags: [roles]
      operationId: createRole
      summary: Create a role
      x-roles: [employer, ops]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Role" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    get:
      tags: [roles]
      operationId: listRoles
      summary: List roles
      description: Talent only ever receives open roles; a `status` filter other than `open` returns an empty list for talent.
      x-roles: [talent, employer, ops]
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/RoleStatus" }
        - name: company
          in: query
          description: Exact match.
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Role" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    role:
      type: apiKey
      in: header
      name: X-Role
      description: A `Persona` (`talent`, `employer` or `ops`). The `mavi_role` cookie is a session fallback.
    actor:
      type: apiKey
      in: header
      name: X-Actor
      description: >-
        The candidate id for talent (required for talent to read anything), an email for ops (required for
        the review operations, whose audit trail records it).
...
  schemas:
...
    CandidateStatus:
      type: string
      enum: [active, archived]
...
    Candidate:
      type: object
      required: [id, full_name, email, phone, location, resume_text, source, status, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        full_name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        location: { type: string, nullable: true }
        resume_text: { type: string }
        source:
          type: string
          nullable: true
          description: Where the record came from; `self` when talent created it.
        status: { $ref: "#/components/schemas/CandidateStatus" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...